refactor(lotto): Test 클래스 분리
